pyproject.toml
[tool.poetry]
name = "animalsgonewild"
version = "1.1.7"
description = "Behaves like the text equivalent of an animal inspired Instagram filter"
authors = ["DSCI 524 - Team 16"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.22.1"
newspaper3k = "^0.2.8"
Pillow = "^9.0.0"
nltk = "^3.6.7"
wordcloud = "^1.8.1"
codecov = "^2.1.12"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
myst-nb = "^0.13.1"
sphinx-autoapi = "^1.8.4"
sphinx-rtd-theme = "^1.0.0"
pytest-cov = "^3.0.0"
python-semantic-release = "^7.24.0"
codecov = "^2.1.12"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_variable = "pyproject.toml:version" # version location
branch = "main"                             # branch to make releases of
changelog_file = "CHANGELOG.md"             # changelog file
build_command = "pip install poetry && poetry build"  # build dists
dist_path = "dist/"                         # where to put dists
upload_to_pypi = false                      # don't auto-upload to PyPI
remove_dist = false                         # don't remove dists
patch_without_tag = "true"                  # always bump version on CD
